fix delta time from float to double

+++ include/Arcade/ECS/ISystemManager.hpp
@@ -40,7 +40,7 @@ namespace Arcade {
                  * @param entityManager The list of entities
                  * @param eventManager The list of events
                  */
-                virtual void update(float deltaTime,
+                virtual void update(double deltaTime,
                 Arcade::ECS::IEventManager &eventManager,
                 Arcade::ECS::IEntityManager &currentEntityManager) = 0;
         };
